auto_merge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto_merge"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
keywords = []
authors = [
  { name = "Leona Maroni", email = "lm@flyingcircus.io" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "gql[requests]",
  "pydantic",
  "pygithub",
  "python-dateutil"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4"
]

[project.scripts]
fc-automerge = "auto_merge.cli:main"


[tool.hatch.version]
path = "src/auto_merge/__about__.py"

[tool.coverage.run]
source_pkgs = ["auto_merge", "tests"]
branch = true
parallel = true
omit = [
  "src/auto_merge/__about__.py",
]

[tool.coverage.paths]
auto_merge = ["src/auto_merge", "*/auto-merge/src/auto_merge"]
tests = ["tests", "*/auto-merge/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
